LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-async-fetch-shared-expected.txt


FAIL Shared CSS module specifier across shadow roots produces the same CSSStyleSheet. assert_equals: First shadow root: expected 1 placeholder(s). expected 1 but got 0
